requirements.txt
Flask==3.0.3
SQLAlchemy==1.4.46
python-dotenv==0.19.2
psycopg2-binary==2.9.9

